requirements.txt
Flask
Flask-CORS
Flask-Migrate
requests
Flask-SQLAlchemy
python-decouple
Werkzeug